fix: clear all 20 mypy errors + 1786 datetime.utcnow deprecation warnings
After this commit, ``make verify`` is fully green:

- ``mypy --strict src/eval_kit/`` — 0 errors (was 20)
- ``pytest -q`` — 554 passed, 6 warnings (was 1786 warnings)
  (the remaining 6 are scipy precision-loss notices unrelated to
  eval-kit, surfaced when permutation tests run on near-identical data)
- ``ruff check src/ tests/`` — clean

## Mypy fixes

Targeted, no blanket ``# type: ignore`` sweep:

- ``contracts/contract.py`` — annotate the temp ``grader`` variable as
  ``Grader`` (the common base) so each branch can reassign to a
  sibling subclass without invariance errors.
- ``metrics/validators.py`` — runtime-check that ``load_class`` returned
  a ``BaseModel`` subclass, then assign through a typed local so
  ``.model_validate`` resolves; raise ``RuntimeError`` otherwise.
- ``execution/registry.py`` — explicit ``cast(type, ...)`` on the
  ``getattr`` return so the declared ``-> type`` is honored.
- ``statistics/pass_at_k.py``, ``statistics/inference.py`` — wrap
  numpy scalar arithmetic in ``float(...)`` to satisfy the declared
  ``-> float`` return type.
- ``statistics/inference.py`` — separate ``perm_diffs_list: list[float]``
  from the later ``np.array`` rebind so the type doesn't change
  mid-function.
- ``calibration/analyzer.py`` — use ``stats.spearmanr(...).statistic``
  (typed) rather than tuple-unpacking the legacy 2-tuple result;
  drop the ``sample_count=`` kwarg from the ``CalibrationResult``
  constructor since it's a computed property, not a field.
- ``baselines/manager.py`` — widen ``compare_to_baseline`` return type
  from ``dict[str, dict[str, Any]]`` to ``dict[str, Any]`` to match
  the ``{"_no_baseline": True}`` sentinel (assertion already covered
  by ``tests/unit/baselines/test_manager.py::test_compare_to_baseline_no_baseline``).
- ``cli/calibrate.py`` (no source change needed) — the analyze callsite
  now passes typecheck because…
- ``calibration/analyzer.py`` — ``analyze`` argument is
  ``Mapping[str, OutcomeLike]`` (covariant in the value) instead of
  the invariant ``dict[...]``, so callers passing
  ``dict[str, Outcome]`` typecheck without coercion.
- ``core/decision_spec.py`` — ``# type: ignore[prop-decorator]`` on the
  two ``@computed_field``+``@property`` stacks; this is mypy's known
  limitation, documented in the pydantic mypy plugin notes.

## Datetime deprecation sweep

Python 3.12 deprecated ``datetime.utcnow()``; Python 3.14 will remove
it. Replace 33 callsites across 11 files (``src/`` + ``tests/``) with a
new ``utc_now()`` helper:

- ``src/eval_kit/core/_time.py`` — internal helper returning
  ``datetime.now(UTC)``.
- All ``Field(default_factory=datetime.utcnow)`` and
  ``datetime.utcnow()`` calls in core models, runner, adapters,
  baselines, conftest, and integration tests now use ``utc_now``.

Behavior change is purely the new datetimes carry tzinfo. No
serialization format change — Pydantic emits the same ISO-8601 string
in either case.

+++ src/eval_kit/statistics/pass_at_k.py
@@ -43,7 +43,7 @@ def pass_at_k(n: int, c: int, k: int) -> float:
 
     # Unbiased estimator: 1 - C(n-c, k) / C(n, k)
     # Equivalent to: 1 - prod((n-c-i)/(n-i) for i in range(k))
-    return 1.0 - np.prod(1.0 - k / np.arange(n - c + 1, n + 1))
+    return float(1.0 - np.prod(1.0 - k / np.arange(n - c + 1, n + 1)))
 
 
 def pass_at_k_estimator(
